Move everything into the correct packages
