+++ examples/src/main/java/dev/rsninja/examples/UsageExample.java
@@ -1,3 +1,5 @@
+package dev.rsninja.examples;
+
 import java.awt.Color;
 
 import game.*;
