+++ examples/src/main/resources/assets/pixelmix.ttf
